chore: use internal sdk for notification template
