+++ templates/bot/ts/notification-function-base/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "echobot",
+  "name": "notificationbot",
   "version": "1.0.0",
-  "description": "Microsoft Teams Toolkit hello world Bot sample",
+  "description": "Microsoft Teams Toolkit Notification Bot Sample",
   "author": "Microsoft",
   "license": "MIT",
   "scripts": {
@@ -21,6 +21,7 @@
   },
   "dependencies": {
     "@microsoft/adaptivecards-tools": "^0.1.3",
+    "@microsoft/teamsfx": "0.6.0-notification.0",
     "botbuilder": "~4.15.0",
     "botbuilder-azure-blobs": "^4.15.0",
     "botbuilder-dialogs": "~4.15.0",
