+++ templates/bot/ts/notification-function-base/src/adaptiveCards/welcome-command.json
@@ -1,0 +1,31 @@
+{
+    "type": "AdaptiveCard",
+    "body": [
+      {
+        "type": "TextBlock",
+        "size": "Medium",
+        "weight": "Bolder",
+        "text": "Your Hello World Bot is Running"
+      },
+      {
+        "type": "TextBlock",
+        "text": "Congratulations! Your hello world bot is running. Click the documentation below to learn more about Bots and the Teams Toolkit.",
+        "wrap": true
+      }
+    ],
+    "actions": [
+      {
+        "type": "Action.OpenUrl",
+        "title": "Bot Framework Docs",
+        "url": "https://docs.microsoft.com/en-us/azure/bot-service/?view=azure-bot-service-4.0"
+      },
+      {
+        "type": "Action.OpenUrl",
+        "title": "Teams Toolkit Docs",
+        "url": "https://aka.ms/teamsfx-docs"
+      }
+    ],
+    "$schema": "http://adaptivecards.io/schemas/adaptive-card.json",
+    "version": "1.4"
+  }
+
