+++ templates/bot/ts/notification-function-base/src/cardModels.ts
@@ -1,0 +1,9 @@
+/**
+ * Adaptive card data model bound to the card template.
+ */
+export interface CardData {
+  title: string;
+  appName: string;
+  description: string;
+  notificationUrl: string;
+}
